requirements.txt
beautifulsoup4==4.5.1
Django==1.10
google==1.9.3
httpsig==1.1.2
nose==1.3.7
numpy==1.11.1
pandas==0.18.1
psycopg2==2.6.2
pycrypto==2.6.1
python-dateutil==2.5.3
python-dwapi==0.1.dev0
python-memcached==1.58
pytz==2016.6.1
requests==2.11.0
six==1.10.0
